refactor: cartItem과 product 인터페이스(타입) 공통화
