+++ src/types/cartItem.ts
@@ -1,8 +1,10 @@
+import { Pageable, Product, Sort } from ".";
+
 export interface GetCartItemsResponse {
   totalElements: number;
   totalPages: number;
   size: number;
-  content: Content[];
+  content: CartItem[];
   number: number;
   sort: Sort;
   pageable: Pageable;
@@ -11,29 +13,9 @@ export interface GetCartItemsResponse {
   last: boolean;
   empty: boolean;
 }
-export interface Content {
+
+export interface CartItem {
   id: number;
   quantity: number;
   product: Product;
-}
-export interface Product {
-  id: number;
-  name: string;
-  price: number;
-  imageUrl: string;
-  category: string;
-  quantity?: number;
-}
-export interface Sort {
-  empty: boolean;
-  sorted: boolean;
-  unsorted: boolean;
-}
-export interface Pageable {
-  offset: number;
-  sort: Sort;
-  paged: boolean;
-  pageNumber: number;
-  pageSize: number;
-  unpaged: boolean;
 }
