+++ src/types/index.ts
@@ -1,2 +1,26 @@
 export type FilterType = "전체" | "식료품" | "패션잡화";
 export type SortType = "높은 가격순" | "낮은 가격순";
+
+export interface Sort {
+  empty: boolean;
+  sorted: boolean;
+  unsorted: boolean;
+}
+
+export interface Pageable {
+  offset: number;
+  sort: Sort;
+  paged: boolean;
+  pageNumber: number;
+  pageSize: number;
+  unpaged: boolean;
+}
+
+export interface Product {
+  id: number;
+  name: string;
+  price: number;
+  imageUrl: string;
+  category: string;
+  quantity?: number;
+}
