+++ src/types/product.ts
@@ -1,5 +1,7 @@
+import { Pageable, Product, Sort } from ".";
+
 export interface GetProductResponse {
-  content: Content[];
+  content: Product[];
   pageable: Pageable;
   totalElements: number;
   totalPages: number;
@@ -10,28 +12,4 @@ export interface GetProductResponse {
   numberOfElements: number;
   first: boolean;
   empty: boolean;
-}
-
-export interface Content {
-  id: number;
-  name: string;
-  price: number;
-  imageUrl: string;
-  category: string;
-  quantity?: number;
-}
-
-export interface Pageable {
-  pageNumber: number;
-  pageSize: number;
-  sort: Sort;
-  offset: number;
-  paged: boolean;
-  unpaged: boolean;
-}
-
-export interface Sort {
-  empty: boolean;
-  sorted: boolean;
-  unsorted: boolean;
 }
